+++ Examples/Effects Example/AsyncImageDemo/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  AsyncImageDemo
+//
+//  Created by Nick Lockwood on 17/10/2011.
+//  Copyright (c) 2011 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (nonatomic, retain) IBOutlet UIWindow *window;
+
+@end